rename announcementbanner to section
