+++ next/src/components/styleguide/showcases/AnnouncementBannerShowcase.tsx
@@ -1,16 +1,16 @@
-import AnnouncementBanner from '@/src/components/common/AnnouncementBanner/AnnouncementBanner'
+import AnnouncementBannerSection from '@/src/components/common/AnnouncementBannerSection/AnnouncementBannerSection'
 
 import Stack from '../Stack'
 import Wrapper from '../Wrapper'
 
-const AnnouncementBannerShowcase = () => {
+const AnnouncementBannerSectionShowcase = () => {
   return (
-    <Wrapper title="Non-critical notification">
+    <Wrapper title="Announcement Banner Section">
       <Stack className="w-304">
-        <AnnouncementBanner />
+        <AnnouncementBannerSection />
       </Stack>
     </Wrapper>
   )
 }
 
-export default AnnouncementBannerShowcase
+export default AnnouncementBannerSectionShowcase
